Split examples route helper logic
